Make HTTPS urls work by default
The POST to Graphite was failing if the URL was HTTPS, unless you
manually set `use_ssl` to true in `graphite_http_options`.

This change populates a reasonable default for `use_ssl` based on the
given `graphite_url` - this appears to have been the behavior added
in #30 but later broken in #31. If someone explicitly sets `use_ssl`,
their setting will override the default.